java/tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
package com.techelevator.tenmo.model;

import java.math.BigDecimal;

public class Account {
    int accountId;
    int userid;
    BigDecimal balance;

    public int getAccountId() {
        return accountId;
    }

    public void setAccountId(int accountId) {
        this.accountId = accountId;
    }

    public int getUserid() {
        return userid;
    }

    public void setUserid(int userid) {
        this.userid = userid;
    }

    public BigDecimal getBalance() {
        return balance;
    }

    public void setBalance(BigDecimal balance) {
        this.balance = balance;
    }
}
